+++ src/main/java/dev/lambdaurora/mcdev/api/LambdaMcDevGradleExtension.java
@@ -12,6 +12,7 @@ import dev.lambdaurora.mcdev.api.mappings.LambdaLayeredMappingsSpecBuilder;
 import org.gradle.api.Action;
 import org.gradle.api.artifacts.ConfigurablePublishArtifact;
 import org.gradle.api.artifacts.Dependency;
+import org.gradle.api.plugins.FeatureSpec;
 import org.gradle.api.provider.Property;
 import org.jetbrains.annotations.NotNull;
 
@@ -36,7 +37,7 @@ public interface LambdaMcDevGradleExtension {
 
 	void setupJarJarCompat();
 
-	MojmapRemapping setupMojmapRemapping();
+	@NotNull MojmapRemapping setupMojmapRemapping();
 
 	default void replaceArtifactInConfiguration(String configuration, Object artifact) {
 		this.replaceArtifactInConfiguration(configuration, artifact, ignored -> {});