+++ src/main/java/dev/lambdaurora/mcdev/api/MappingVariant.java
@@ -1,0 +1,12 @@
+package dev.lambdaurora.mcdev.api;
+
+import org.gradle.api.Named;
+import org.gradle.api.attributes.Attribute;
+
+public interface MappingVariant extends Named {
+	Attribute<MappingVariant> ATTRIBUTE = Attribute.of("net.minecraft.mappings", MappingVariant.class);
+
+	String INTERMEDIARY = "intermediary";
+
+	String MOJMAP = "mojmap";
+}