+++ src/main/java/dev/lambdaurora/mcdev/ext/LambdaMcDevGradleExtensionImpl.java
@@ -9,6 +9,7 @@
 package dev.lambdaurora.mcdev.ext;
 
 import dev.lambdaurora.mcdev.api.LambdaMcDevGradleExtension;
+import dev.lambdaurora.mcdev.api.MappingVariant;
 import dev.lambdaurora.mcdev.api.ModManifests;
 import dev.lambdaurora.mcdev.api.MojmapRemapping;
 import dev.lambdaurora.mcdev.api.manifest.Fmj;
@@ -18,12 +19,19 @@ import dev.lambdaurora.mcdev.mappings.LambdaLayeredMappingsSpecBuilderImpl;
 import dev.lambdaurora.mcdev.task.GenerateFmjTask;
 import dev.lambdaurora.mcdev.task.GenerateNeoForgeJiJDataTask;
 import dev.lambdaurora.mcdev.task.GenerateNmtTask;
+import dev.lambdaurora.mcdev.util.Utils;
 import dev.yumi.commons.function.YumiPredicates;
 import net.fabricmc.loom.api.LoomGradleExtensionAPI;
+import net.fabricmc.loom.build.nesting.NestableJarGenerationTask;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.ConfigurablePublishArtifact;
 import org.gradle.api.artifacts.Dependency;
+import org.gradle.api.artifacts.ModuleDependency;
+import org.gradle.api.attributes.*;
+import org.gradle.api.attributes.java.TargetJvmVersion;
+import org.gradle.api.component.AdhocComponentWithVariants;
+import org.gradle.api.component.ConfigurationVariantDetails;
 import org.gradle.api.plugins.JavaPluginExtension;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.Provider;
@@ -33,12 +41,13 @@ import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
 import java.nio.file.Path;
+import java.util.ArrayList;
 
 public class LambdaMcDevGradleExtensionImpl implements LambdaMcDevGradleExtension {
-	private final Project project;
+	final Project project;
 	private final Property<String> namespace;
 	private final JavaPluginExtension java;
-	private final LoomGradleExtensionAPI loom;
+	final LoomGradleExtensionAPI loom;
 	private final Manifests manifests;
 
 	public LambdaMcDevGradleExtensionImpl(
@@ -105,24 +114,130 @@ public class LambdaMcDevGradleExtensionImpl implements LambdaMcDevGradleExtensio
 	}
 
 	@Override
-	public MojmapRemapping setupMojmapRemapping() {
-		var mojmap = this.java.getSourceSets().create("mojmap");
-		var config = this.project.getConfigurations().create("mojangMappings");
+	public @NotNull MojmapRemapping setupMojmapRemapping() {
+		final var configurations = this.project.getConfigurations();
+		final var tasks = this.project.getTasks();
+		final var objects = this.project.getObjects();
+		final boolean hasSources = tasks.findByPath("remapSourcesJar") != null;
 
+		var mojmapConfig = configurations.create("mojangMappings");
 		this.project.getDependencies().add("mojangMappings", this.loom.officialMojangMappings());
 
-		this.java.registerFeature("mojmap", feature -> {
-			feature.usingSourceSet(mojmap);
-			feature.withSourcesJar();
+		var published = this.java.getSourceSets().create("mojmap");
 
-			this.project.afterEvaluate(project -> {
-				var configurations = project.getConfigurations();
-				configurations.getByName("mojmapApiElements").extendsFrom(configurations.getByName("apiElements"));
-				configurations.getByName("mojmapRuntimeElements").extendsFrom(configurations.getByName("runtimeElements"));
+		configurations.getByName("apiElements").attributes(this.setMappingsAttribute(MappingVariant.INTERMEDIARY));
+		configurations.getByName("runtimeElements").attributes(this.setMappingsAttribute(MappingVariant.INTERMEDIARY));
+
+		if (hasSources) {
+			configurations.getByName("sourcesElements").attributes(this.setMappingsAttribute(MappingVariant.INTERMEDIARY));
+		}
+
+		final var apiConfig = configurations.create(published.getApiConfigurationName(), config -> {
+			config.attributes(this.setMappingsAttribute(MappingVariant.MOJMAP));
+		});
+
+		final var apiElements = configurations.create(published.getApiElementsConfigurationName(), config -> {
+			config.extendsFrom(apiConfig);
+			config.attributes(attributes -> {
+				this.setMappingsAttribute(attributes, MappingVariant.MOJMAP);
+				attributes.attribute(Category.CATEGORY_ATTRIBUTE, objects.named(Category.class, Category.LIBRARY));
+				attributes.attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements.class, LibraryElements.JAR));
+				attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage.class, Usage.JAVA_API));
+				attributes.attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling.class, Bundling.EXTERNAL));
+				var languageVersion = this.java.getToolchain().getLanguageVersion();
+				if (languageVersion.isPresent()) {
+					attributes.attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, languageVersion.get().asInt());
+				}
+			});
+		});
+		final var runtimeElements = configurations.create(published.getRuntimeElementsConfigurationName(), config -> {
+			config.extendsFrom(apiConfig, configurations.getByName(published.getImplementationConfigurationName()));
+			config.attributes(attributes -> {
+				this.setMappingsAttribute(attributes, MappingVariant.MOJMAP);
+				attributes.attribute(Category.CATEGORY_ATTRIBUTE, objects.named(Category.class, Category.LIBRARY));
+				attributes.attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements.class, LibraryElements.JAR));
+				attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage.class, Usage.JAVA_RUNTIME));
+				attributes.attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling.class, Bundling.EXTERNAL));
+				var languageVersion = this.java.getToolchain().getLanguageVersion();
+				if (languageVersion.isPresent()) {
+					attributes.attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, languageVersion.get().asInt());
+				}
+			});
+		});
+		final var sourcesElements = configurations.create(published.getSourcesElementsConfigurationName(), config -> {
+			config.attributes(attributes -> {
+				this.setMappingsAttribute(attributes, MappingVariant.MOJMAP);
+				attributes.attribute(Category.CATEGORY_ATTRIBUTE, objects.named(Category.class, Category.DOCUMENTATION));
+				attributes.attribute(DocsType.DOCS_TYPE_ATTRIBUTE, objects.named(DocsType.class, DocsType.SOURCES));
+				attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage.class, Usage.JAVA_RUNTIME));
 			});
 		});
 
-		return new MojmapRemappingImpl(this, mojmap, config);
+		final var include = configurations.register(published.getName() + "Include", config -> {
+			config.setCanBeConsumed(false);
+			config.setCanBeResolved(false);
+		});
+		final var includeInternal = configurations.register(published.getName() + "IncludeInternal", config -> {
+			config.getDependencies().addAllLater(this.project.provider(() -> {
+				var dependencies = new ArrayList<Dependency>();
+
+				for (Dependency dependency : include.get().getIncoming().getDependencies()) {
+					if (dependency instanceof HasConfigurableAttributes<?> hasAttributes) {
+						Category category = hasAttributes.getAttributes().getAttribute(Category.CATEGORY_ATTRIBUTE);
+
+						if (category != null && (category.getName().equals(Category.ENFORCED_PLATFORM) || category.getName().equals(Category.REGULAR_PLATFORM))) {
+							dependencies.add(dependency);
+							continue;
+						} else if (dependency instanceof ModuleDependency moduleDependency) {
+							var copy = moduleDependency.copy();
+							copy.setTransitive(false);
+							dependencies.add(copy);
+							continue;
+						}
+					}
+
+					dependencies.add(dependency);
+				}
+
+				return dependencies;
+			}));
+			config.attributes(attributes -> {
+				this.setMappingsAttribute(attributes, MappingVariant.MOJMAP);
+				attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage.class, Usage.JAVA_RUNTIME));
+				attributes.attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements.class, LibraryElements.JAR));
+				attributes.attribute(Category.CATEGORY_ATTRIBUTE, objects.named(Category.class, Category.LIBRARY));
+				attributes.attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling.class, Bundling.EXTERNAL));
+			});
+		});
+
+		var processIncludeJars = tasks.register("process" + Utils.capitalize(published.getName()) + "IncludeJars", NestableJarGenerationTask.class, task -> {
+			task.from(includeInternal.get());
+			task.getOutputDirectory().set(project.getLayout().getBuildDirectory().dir(task.getName()));
+		});
+
+		final var javaComponent = ((AdhocComponentWithVariants) this.project.getComponents().getByName("java"));
+		javaComponent.addVariantsFromConfiguration(apiElements, details -> {
+			details.mapToMavenScope("compile");
+			details.mapToOptional();
+		});
+		javaComponent.addVariantsFromConfiguration(runtimeElements, details -> {
+			details.mapToMavenScope("runtime");
+			details.mapToOptional();
+		});
+		javaComponent.addVariantsFromConfiguration(sourcesElements, ConfigurationVariantDetails::mapToOptional);
+
+		return new MojmapRemappingImpl(this, published, mojmapConfig, processIncludeJars);
+	}
+
+	private Action<AttributeContainer> setMappingsAttribute(String variant) {
+		return attributes -> this.setMappingsAttribute(attributes, variant);
+	}
+
+	private void setMappingsAttribute(AttributeContainer attributes, String variant) {
+		attributes.attribute(
+				MappingVariant.ATTRIBUTE,
+				this.project.getObjects().named(MappingVariant.class, variant)
+		);
 	}
 
 	@Override