+++ src/main/java/dev/lambdaurora/mcdev/util/Utils.java
@@ -1,0 +1,7 @@
+package dev.lambdaurora.mcdev.util;
+
+public final class Utils {
+	public static String capitalize(String str) {
+		return str.substring(0, 1).toUpperCase() + str.substring(1);
+	}
+}